# api/docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.CreateProduct:
    properties:
      barcode:
        type: string
      category_id:
        type: string
      photo:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
...
  models.Product:
    properties:
      barcode:
        type: string
      category_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      photo:
        type: string
      price:
        type: number
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateProduct:
    properties:
      barcode:
        type: string
      category_id:
        type: string
      id:
        type: string
      photo:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/product:
    get:
      consumes:
      - application/json
      description: Get a list of products with optional filtering.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Number of items to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of items to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a list of products
      tags:
      - product
    post:
      consumes:
      - application/json
      description: Create a new product in the market system.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.CreateProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created product
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new product
      tags:
      - product
  /v1/product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing product.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a product
      tags:
      - product
    get:
      consumes:
      - application/json
      description: Get product details by its ID.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a product by ID
      tags:
      - product
    put:
      consumes:
      - application/json
      description: Update an existing product.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProduct'
      produces:
      - application/json
      responses:
        "202":
          description: Updated product
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update a product
      tags:
      - product
...
swagger: "2.0"